update hosts and mappings
put the dashboard host in linkerd-viz namespace

disable almost all the hard parts in demo-magic

We are just going to assume all of the hard parts are done and linkerd
is online, so is Flagger, all necessary pods have been checked and are
meshed, etc. and ambassador stack is ready to go with a load balancer
that has been exposed to the public internet for demo users to hit while
the interactive demo is ongoing.

(This does not interfere with any of our existing ingress magic, except
that since we only have one public IP, only one of ingress-public and
ambassador can be made public at a given time. This means if we want the
webhook to work, we will need to add a Mapping and Host so that it can
work through ambassador.)